Python/Exercicios/Certificado/requirements.txt
deprecation==2.1.0
img2pdf==0.4.4
lxml==4.9.1
packaging==21.3
pikepdf==6.2.1
Pillow==9.3.0
pyparsing==3.0.9
